rewrite an algorithm
