+++ src/js/default.json
@@ -1,0 +1,23 @@
+{
+  "1": {
+    "name": "1"
+  },
+  "2": {
+    "name": "2"
+  },
+  "3": {
+    "name": "3"
+  },
+  "4": {
+    "name": "4"
+  },
+  "5": {
+    "name": "5"
+  },
+  "6": {
+    "name": "6"
+  },
+  "7": {
+    "name": "7"
+  }
+}
